block digging works completely properly now
though there may be some edge cases, particularly when > 0 net latency
gets involved

+++ core/src/com/ore/infinium/OreWorld.java
@@ -907,7 +907,7 @@ public class OreWorld {
         /**
          * max starting health of the block
          */
-        public short blockTotalHealth;
+        public float blockTotalHealth;
 
         //if this type is a type of ore (like stone, copper, ...)
         public enum BlockCategory {
